Changed the interactive decoder to work with integer bin definitions. For the PicoHarp and HydraHarp, the limiting resolution is 4ps and 1ps, respectively, so all bin definitions are true integer values. For the TimeHarp the values are divisions of 1us or so, so these are handled as double-precision floats but rounded to the nearest ps when needed.

+++ src/interactive.h
@@ -38,8 +38,8 @@
 
 typedef struct {
 	uint32_t curve;
-	float64_t bin_left;
-	float64_t bin_right;
+	int64_t bin_left;
+	int64_t bin_right;
 	uint32_t counts;
 } pq_interactive_bin_t;
